+++ zotero-plugin.config.ts
@@ -1,16 +1,20 @@
 import { defineConfig } from "zotero-plugin-scaffold";
 import pkg from "./package.json";
 
+const xpiName = `${pkg.config.addonName} V${pkg.version}`;
+const encodedXpiName = encodeURIComponent(xpiName);
+
 export default defineConfig({
   source: ["src", "addon"],
   dist: ".scaffold/build",
   name: pkg.config.addonName,
   id: pkg.config.addonID,
   namespace: pkg.config.addonRef,
+  xpiName,
   updateURL: `https://github.com/{{owner}}/{{repo}}/releases/download/release/${
     pkg.version.includes("-") ? "update-beta.json" : "update.json"
   }`,
-  xpiDownloadLink: "https://github.com/{{owner}}/{{repo}}/releases/download/v{{version}}/{{xpiName}}.xpi",
+  xpiDownloadLink: `https://github.com/{{owner}}/{{repo}}/releases/download/v{{version}}/${encodedXpiName}.xpi`,
 
   build: {
     assets: ["addon/**/*.*"],